+++ src/main/java/net/sourceforge/jeuclid/element/MathSup.java
@@ -61,12 +61,12 @@ public class MathSup extends AbstractMathElementWithSubSuper implements
      *            The position of the baseline.
      */
     @Override
-    public void paint(final Graphics2D g, final int posX, final int posY) {
+    public void paint(final Graphics2D g, final float posX, final float posY) {
         super.paint(g, posX, posY);
         final MathElement e1 = this.getMathElement(0);
         final MathElement e2 = this.getMathElement(1);
 
-        final int posY2 = posY - this.getSuperBaseLineShift(g);
+        final float posY2 = posY - this.getSuperBaseLineShift(g);
         // final int posY1 = posY + e1DescentHeight + e2.getAscentHeight(g)
         // - middleshift;
         //
@@ -82,20 +82,20 @@ public class MathSup extends AbstractMathElementWithSubSuper implements
 
     /** {@inheritDoc} */
     @Override
-    public int calculateWidth(final Graphics2D g) {
+    public float calculateWidth(final Graphics2D g) {
         return this.getMathElement(0).getWidth(g)
                 + this.getMathElement(1).getWidth(g);
     }
 
     /** {@inheritDoc} */
     @Override
-    public int calculateAscentHeight(final Graphics2D g) {
+    public float calculateAscentHeight(final Graphics2D g) {
         return this.caclulateAscentHeightWithSuper(g);
     }
 
     /** {@inheritDoc} */
     @Override
-    public int calculateDescentHeight(final Graphics2D g) {
+    public float calculateDescentHeight(final Graphics2D g) {
         return this.getMathElement(0).getDescentHeight(g);
     }
