+++ src/main/java/net/sourceforge/jeuclid/element/generic/AbstractElementWithDelegates.java
@@ -65,25 +65,25 @@ public abstract class AbstractElementWithDelegates extends
 
     /** {@inheritDoc} */
     @Override
-    public int calculateAscentHeight(final Graphics2D g) {
+    public float calculateAscentHeight(final Graphics2D g) {
         return ElementListSupport.getAscentHeight(g, this.delegates);
     }
 
     /** {@inheritDoc} */
     @Override
-    public int calculateDescentHeight(final Graphics2D g) {
+    public float calculateDescentHeight(final Graphics2D g) {
         return ElementListSupport.getDescentHeight(g, this.delegates);
     }
 
     /** {@inheritDoc} */
     @Override
-    public int calculateWidth(final Graphics2D g) {
+    public float calculateWidth(final Graphics2D g) {
         return ElementListSupport.getWidth(g, this.delegates);
     }
 
     /** {@inheritDoc} */
     @Override
-    public void paint(final Graphics2D g, final int posX, final int posY) {
+    public void paint(final Graphics2D g, final float posX, final float posY) {
         super.paint(g, posX, posY);
         ElementListSupport.paint(g, posX, posY, this.delegates);
     }
